refactor(US-2650) — extraire buildTreatmentView vers src/lib/insulin/ (dette revue #658)
Ferme le finding LOW de la revue de #658 : une page (patient) importait buildTreatmentView
depuis un helper co-localisé (dashboard)/patients/[id]/ (couplage cross-route-group).

- git mv (dashboard)/patients/[id]/treatment-view.ts → src/lib/insulin/treatment-view.ts.
  Module PUR (imports @/-absolus, aucune dépendance RSC/Prisma) → portable tel quel.
- 5 importeurs mis à jour → @/lib/insulin/treatment-view (build-patient-record, page patient,
  review/page, ReviewClient, test).

Zéro changement de comportement. Vérifs : tsc · eslint · anti-drift 267 · 3678 tests.

Refs #633 (US-2650).
