+++ src/app/(dashboard)/patients/[id]/review/ReviewClient.tsx
@@ -28,7 +28,7 @@ import { CgmChart } from "@/components/diabeo/CgmChart"
 import { bcp47 } from "@/i18n/config"
 import { PROPOSAL_MAJOR_CHANGE_PCT } from "@/lib/review-constants"
 import type { GlycemiaView } from "../glycemia-view"
-import type { TreatmentView } from "../treatment-view"
+import type { TreatmentView } from "@/lib/insulin/treatment-view"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Button } from "@/components/ui/button"
